backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
google-generativeai
python-multipart
httpx
google-auth
requests
pydantic[email]
easyocr
torch
torchvision
trafilatura
newspaper3k
lxml[html_clean]
pillow
aiofiles
reportlab
svglib
groq
